src/webhook-handler/requirements.txt
Flask==3.0.3
google-cloud-pubsub==2.21.5
google-cloud-secret-manager==2.20.2
gunicorn==22.0.0
